nutri-health-api/requirements.txt
# FastAPI and server
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# File handling
python-multipart==0.0.12
pillow==11.0.0

# AI Integration
google-generativeai==0.8.3

# Environment variables
python-dotenv==1.0.1

# Utilities
pydantic==2.10.1
pydantic-settings==2.6.1
